fastapi-ariadne/pyproject.toml
[tool.poetry]
name = "fastapi-ariadne"
version = "0.1.0"
description = ""
authors = ["Matthew Chung <matthewchung74@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.63.0"
uvicorn = {extras = ["standard"], version = "^0.13.4"}
websockets = "^8.1"
databases = {extras = ["postgresql"], version = "^0.4.1"}
async-exit-stack = "^1.0.1"
async_generator = "^1.10"
ariadne = "^0.12.0"
psycopg2-binary = "^2.8.6"
python-jose = "^3.2.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
httpx = "^0.17.0"
pytest-asyncio = "^0.14.0"
python-dotenv = "^0.15.0"
alembic = "^1.5.5"

[tool.poetry.dev-dependencies]
pytest = "^6.2.2"
python-graphql-client = "^0.4.2"
pywsitest = "^0.3.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
